chore(project): rename is-input to isInput
